+++ pmwdlib/pybind11_kernel_helpers.h
@@ -26,7 +26,7 @@ pybind11::capsule encapsulate_function(T* fn) {
 template <typename T>
 pybind11::bytes build_descriptor(int64_t n_particle, T ptcl_spacing, T cell_size,
                                  T offset_1, T offset_2, T offset_3,
-                                 int64_t stride_1, int64_t stride_2, int64_t stride_3) {
+                                 uint32_t stride_1, uint32_t stride_2, uint32_t stride_3) {
   return pack_descriptor(
       PmwdDescriptor<T>{n_particle, ptcl_spacing, cell_size, {offset_1, offset_2, offset_3}, {stride_1, stride_2, stride_3}});
 }
